+++ test/test_eol.py
@@ -18,6 +18,20 @@ class DocTestsTestCase(unittest.TestCase):
         test = doctest.DocFileTest("api.doctests")
         test.runTest()
 
+    if sys.version_info[0] == 2:
+        def test_api2(self):
+            """API tests for Python 2 syntax"""
+            if sys.version_info[:2] < (2,4):
+                raise TestSkipped("no DocFileTest in Python <=2.3")
+            test = doctest.DocFileTest("api2.doctests")
+            test.runTest()
+    
+    if sys.version_info[0] > 2:
+        def test_api3(self):
+            """API tests for Python 3 syntax"""
+            test = doctest.DocFileTest("api3.doctests")
+            test.runTest()
+
     def test_internal(self):
         import eol
         doctest.testmod(eol)